[#2080,#2154,irods/irods#7265] Minor reorganization + CMake TLC
- Project tree structure reorganized to match that of main irods repo.
- Removed unused read_shared_memory.cpp.
- Fixed some system #includes not having angle brackets.
- Added quotes around paths in CMake files.
- Changed all instances of CMAKE_SOURCE_DIR to CMAKE_CURRENT_SOURCE_DIR.
- Removed CMAKE_SHARED_LINKER_FLAGS_INIT and
  CMAKE_SHARED_LINKER_FLAGS_RELEASE_INIT, as no shared libraries are built.
- Added IRODS_PLUGINS_DIRECTORY to
  CPACK_RPM_EXCLUDE_FROM_AUTO_FILELIST_ADDITION.
- Added CMAKE_EXE_LINKER_FLAGS_INIT and CMAKE_EXE_LINKER_FLAGS_RELEASE_INIT
  for unit tests.
- [#2080] Rewrote unit test cmake to match that of main irods repo.
- [#2154] Replaced curl-devel package dependency declaration with libcurl
