+++ s3_resource/include/irods/private/s3_resource/s3_operations.hpp
@@ -1,5 +1,5 @@
-#ifndef _IRODS_S3_CACHELESS_OPERATIONS_H
-#define _IRODS_S3_CACHELESS_OPERATIONS_H
+#ifndef IRODS_S3_RESOURCE_OPERATIONS_HPP
+#define IRODS_S3_RESOURCE_OPERATIONS_HPP
 
 // =-=-=-=-=-=-=-
 // irods includes
