+++ s3_transport/include/irods/private/s3_transport/callbacks.hpp
@@ -30,11 +30,11 @@
 #include <boost/filesystem.hpp>
 
 // local includes
-#include "circular_buffer.hpp"
-#include "managed_shared_memory_object.hpp"
-#include "s3_multipart_shared_data.hpp"
-#include "s3_transport_types.hpp"
-#include "s3_transport_logging_category.hpp"
+#include "irods/private/s3_transport/circular_buffer.hpp"
+#include "irods/private/s3_transport/managed_shared_memory_object.hpp"
+#include "irods/private/s3_transport/multipart_shared_data.hpp"
+#include "irods/private/s3_transport/types.hpp"
+#include "irods/private/s3_transport/logging_category.hpp"
 
 namespace irods::experimental::io::s3_transport
 {
