+++ s3_transport/include/irods/private/s3_transport/lock_and_wait_strategy.hpp
