+++ unit_tests/cmake/utils.cmake
@@ -7,4 +7,5 @@ function(unset_irods_test_variables)
     unset(IRODS_TEST_SOURCE_FILES)
     unset(IRODS_TEST_INCLUDE_PATH)
     unset(IRODS_TEST_LINK_LIBRARIES)
+    unset(IRODS_TEST_LINK_OBJLIBRARIES)
 endfunction()
